Exception Handling

+++ src/main/java/com/iiitb/giftcartdevops/customer/CustomerController.java
@@ -4,6 +4,7 @@ import com.iiitb.giftcartdevops.address.Address;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.xml.ws.http.HTTPException;
 import java.util.List;
 import java.util.Optional;
 
@@ -37,21 +38,39 @@ public class CustomerController {
     }
 
     @RequestMapping(method = RequestMethod.POST, value = "/customer")
-        public void addCustomerByID(@RequestBody Customer customer)
-        {
-            customerService.addCustomer(customer);
+        public void addCustomerByID(@RequestBody Customer customer) throws Exception {
+            Customer check = customerService.getCustomerByEmail(customer.email);
+            if(check==null) {
+                customerService.addCustomer(customer);
+            }
+            else
+            {
+                throw new Exception("Username exists");
+            }
         }
 
     @RequestMapping(method = RequestMethod.PUT, value = "/customer/{id}")
-    public void updateCustomerByID(@RequestBody Customer customer, @PathVariable Integer id)
-    {
-        customerService.updateCustomer(id,customer);
+    public void updateCustomerByID(@RequestBody Customer customer, @PathVariable Integer id) throws Exception {
+        Optional<Customer> check = customerService.getCustomer(id);
+        if(check==null) {
+            throw new Exception("Username does not exist");
+        }
+        else
+        {
+            customerService.updateCustomer(id,customer);
+        }
+
     }
 
 
     @RequestMapping(method = RequestMethod.DELETE, value = "/customer/{id}")
-    public void deleteCustomerByID(@PathVariable Integer id)
-    {
+    public void deleteCustomerByID(@PathVariable Integer id) throws Exception {
+        Optional<Customer> check = customerService.getCustomer(id);
+        if(check == null)
+        {
+            throw new Exception("User does not exist");
+        }
+        else
         customerService.deleteCustomer(id);
     }
 }
